+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Request URL for sending analytics data [RGA-936]
 - Display selected item in playback menu [RGOeX-25733]
+- Pause on mouse hover or focus on the transcript item/block [RGOeX-26175]
 
 ## [1.3.0] 2024-03-12